requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dotenv==1.0.0
matplotlib==3.8.2